requirements.txt
streamlit==1.24.0
tensorflow-cpu==2.13.0
numpy==1.23.5
Pillow==9.5.0
opencv-python==4.7.0
huggingface_hub==0.16.0